+++ lib/redux_machine.dart
@@ -1,7 +1,7 @@
 // Copyright (c) 2017, Anatoly Pulyaevskiy. All rights reserved. Use of this source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
-/// State machine which uses Redux flow.
+/// Redux Store implementation which can also be used as a state machine.
 library redux_machine;
 
 export 'src/store.dart';